Update to latest protos (#426)
## tl;dr

- Updates to the latest protos
- Handles rename of "address" fields to "identifier"

## Notes

There is still some deferred work here for properly storing the identifier type. The backend assumes all identifiers are Ethereum addresses even when they are passkeys. We should store the distinction in the DB, and support reverse-resolution of passkey -> inbox_id.

When we add that support, all passkeys registered before are going to get misplaced since they are being treated as Ethereum addresses. Maybe we can do some sort of migration based on length or prefix to find any passkeys in the DB?