More performant DB migration for wrapper_algorithm column (#455)
```
UPDATE
	welcome_messages
SET
	wrapper_algorithm = 0;
```

Updates every row in the DB, and times out on the dev network. The alternative in this PR allows Postgres to internally optimize away the writes (runs instantly).

I've manually applied the migration on the dev network, and the migration is marked as applied on dev and prod, so landing this PR shouldn't re-run the migration or change anything in practice.